… plugins/catalog-backend-module-msgraph/api-report.md
… plugins/catalog-backend-module-msgraph/src/microsoftGraph/client.test.ts
… plugins/catalog-backend-module-msgraph/src/microsoftGraph/read.test.ts