Fixes #14757

## Hey, I just made a Pull Request!

Improved performance by reducing the number of calls being made to the Microsoft Graph API.
* Increased the number of rows returned at a time from 100 to 999.
* Load user data in batches when getting users by group membership, rather than making individual calls per user.

In my local environment, this improved the time for ingesting 42 users and 28 groups from 173 seconds to 113 (a roughly 33% improvement). This will show biggest improvements in local environments where the latency of extra calls is more expensive, but should show some benefits in production environments.

Added support for natively loading only users or only groups, without having hack it in (e.g. by using queries that always return false). If you don't add any configuration for groups or users, then they won't be ingested.

Improved Logging to provide more indications that something is happening throughout the ingestion process.  This should hopefully help with the questions we get around the Microsoft Graph plugin not loading users, when the issue is not waiting long enough.  As an example of hte new logging:

Before

```plain
[1] 2023-01-09T20:07:44.063Z catalog info Reading msgraph users and groups type=plugin class=MicrosoftGraphOrgEntityProvider taskId=MicrosoftGraphOrgEntityProvider:default:refresh taskInstanceId=4bd2db5e-943c-440d-932a-b3cddf770e5e
[1] 2023-01-09T20:08:13.857Z catalog info groupMemberUsers 43 type=plugin class=MicrosoftGraphOrgEntityProvider taskId=MicrosoftGraphOrgEntityProvider:default:refresh taskInstanceId=4bd2db5e-943c-440d-932a-b3cddf770e5e
[0] webpack compiled successfully
[1] 2023-01-09T20:10:37.077Z catalog info Read 42 msgraph users and 28 msgraph groups in 173.0 seconds. Committing... type=plugin class=MicrosoftGraphOrgEntityProvider taskId=MicrosoftGraphOrgEntityProvider:default:refresh taskInstanceId=4bd2db5e-943c-440d-932a-b3cddf770e5e
[1] 2023-01-09T20:10:37.139Z catalog info Committed 42 msgraph users and 28 msgraph groups in 0.1 seconds. type=plugin class=MicrosoftGraphOrgEntityProvider taskId=MicrosoftGraphOrgEntityProvider:default:refresh taskInstanceId=4bd2db5e-943c-440d-932a-b3cddf770e5e
```

After

```plain
[1] 2023-01-09T20:33:32.844Z catalog info Reading Microsoft graph users and groups type=plugin class=MicrosoftGraphOrgEntityProvider taskId=MicrosoftGraphOrgEntityProvider:default:refresh taskInstanceId=94390bd8-dd4d-4ecc-8cc6-0822aaa8b9c7
[1] 2023-01-09T20:33:32.852Z catalog info Loading users by group membership type=plugin class=MicrosoftGraphOrgEntityProvider taskId=MicrosoftGraphOrgEntityProvider:default:refresh taskInstanceId=94390bd8-dd4d-4ecc-8cc6-0822aaa8b9c7
[1] 2023-01-09T20:33:46.318Z catalog info Read users from group type=plugin class=MicrosoftGraphOrgEntityProvider taskId=MicrosoftGraphOrgEntityProvider:default:refresh taskInstanceId=94390bd8-dd4d-4ecc-8cc6-0822aaa8b9c7 groupId=REDACTED groupName=REDACTED memberCount=5
[1] 2023-01-09T20:33:46.354Z catalog info Read users from group type=plugin class=MicrosoftGraphOrgEntityProvider taskId=MicrosoftGraphOrgEntityProvider:default:refresh taskInstanceId=94390bd8-dd4d-4ecc-8cc6-0822aaa8b9c7 groupId=REDACTED groupName=REDACTED memberCount=0
...
[1] 2023-01-09T20:34:02.150Z catalog info Read users from group membership type=plugin class=MicrosoftGraphOrgEntityProvider taskId=MicrosoftGraphOrgEntityProvider:default:refresh taskInstanceId=94390bd8-dd4d-4ecc-8cc6-0822aaa8b9c7 groupCount=27 userCount=43
[1] 2023-01-09T20:34:02.151Z catalog info Transforming users type=plugin class=MicrosoftGraphOrgEntityProvider taskId=MicrosoftGraphOrgEntityProvider:default:refresh taskInstanceId=94390bd8-dd4d-4ecc-8cc6-0822aaa8b9c7
[1] 2023-01-09T20:34:59.023Z catalog info Finished transforming users type=plugin class=MicrosoftGraphOrgEntityProvider taskId=MicrosoftGraphOrgEntityProvider:default:refresh taskInstanceId=94390bd8-dd4d-4ecc-8cc6-0822aaa8b9c7 microsoftUserCount=43 backstageUserCount=42
[1] 2023-01-09T20:34:59.024Z catalog info Loading Organization type=plugin class=MicrosoftGraphOrgEntityProvider taskId=MicrosoftGraphOrgEntityProvider:default:refresh taskInstanceId=94390bd8-dd4d-4ecc-8cc6-0822aaa8b9c7
[1] 2023-01-09T20:35:03.086Z catalog info Loading Groups type=plugin class=MicrosoftGraphOrgEntityProvider taskId=MicrosoftGraphOrgEntityProvider:default:refresh taskInstanceId=94390bd8-dd4d-4ecc-8cc6-0822aaa8b9c7
[0] webpack compiled successfully
[1] 2023-01-09T20:35:26.670Z catalog info Loaded Groups type=plugin class=MicrosoftGraphOrgEntityProvider taskId=MicrosoftGraphOrgEntityProvider:default:refresh taskInstanceId=94390bd8-dd4d-4ecc-8cc6-0822aaa8b9c7 count=28
[1] 2023-01-09T20:35:26.671Z catalog info Resolving relations type=plugin class=MicrosoftGraphOrgEntityProvider taskId=MicrosoftGraphOrgEntityProvider:default:refresh taskInstanceId=94390bd8-dd4d-4ecc-8cc6-0822aaa8b9c7
[1] 2023-01-09T20:35:26.680Z catalog info Read 42 users and 28 groups from Microsoft graph in 113.8 seconds. Committing... type=plugin class=MicrosoftGraphOrgEntityProvider taskId=MicrosoftGraphOrgEntityProvider:default:refresh taskInstanceId=94390bd8-dd4d-4ecc-8cc6-0822aaa8b9c7
[1] 2023-01-09T20:35:26.712Z catalog info Committed 42 users and 28 groups from Microsoft graph in 0.0 seconds. type=plugin class=MicrosoftGraphOrgEntityProvider taskId=MicrosoftGraphOrgEntityProvider:default:refresh taskInstanceId=94390bd8-dd4d-4ecc-8cc6-0822aaa8b9c7
```

Other misc changes

* Moved default transformers out to their own file
* Reduced duplication in tests by reusing examples, rather than duplicate the same sample data in multiple tests.

#### :heavy_check_mark: Checklist

<!--- Please include the following in your Pull Request when applicable: -->

- [X] A changeset describing the change and affected packages. ([more info](https://github.com/backstage/backstage/blob/master/CONTRIBUTING.md#creating-changesets))
- [ ] Added or updated documentation
- [X] Tests for new functionality and regression tests for bug fixes
- [ ] Screenshots attached (for UI changes)
- [X] All your commits have a `Signed-off-by` line in the message. ([more info](https://github.com/backstage/backstage/blob/master/CONTRIBUTING.md#developer-certificate-of-origin))